+++ tests/framework/results/test_stratification.py
@@ -45,7 +45,7 @@ def test_stratification(mapper, is_vectorized):
         mapper=mapper,
         is_vectorized=is_vectorized,
     )
-    output = my_stratification(STUDENT_TABLE)
+    output = my_stratification.stratify(STUDENT_TABLE)
     assert output.eq(STUDENT_HOUSES).all()
 
 
@@ -56,13 +56,19 @@ def test_stratification(mapper, is_vectorized):
             [],
             HOUSE_CATEGORIES,
             None,
-            f"No mapper provided for stratification {NAME} with 0 stratification sources.",
+            (
+                f"No mapper but 0 stratification sources are provided for stratification {NAME}. "
+                "The list of sources must be of length 1 if no mapper is provided."
+            ),
         ),
         (
             NAME_COLUMNS,
             HOUSE_CATEGORIES,
             None,
-            f"No mapper provided for stratification {NAME} with {len(NAME_COLUMNS)} stratification sources.",
+            (
+                f"No mapper but {len(NAME_COLUMNS)} stratification sources are provided for stratification {NAME}. "
+                "The list of sources must be of length 1 if no mapper is provided."
+            ),
         ),
         (
             [],
@@ -143,7 +149,7 @@ def test_stratification_call_raises(
         NAME, sources, HOUSE_CATEGORIES, [], mapper, is_vectorized
     )
     with pytest.raises(expected_exception, match=re.escape(error_match)):
-        my_stratification(STUDENT_TABLE)
+        my_stratification.stratify(STUDENT_TABLE)
 
 
 @pytest.mark.parametrize("default_stratifications", [["age", "sex"], ["age"], []])